Upstream some-selected-bugfixes-from-cinelera-cv-until-Jan-2018
Went over the accummulated diff towards current cinelerra-cv master
and picked a lot of minor bugfixes without any functionality changes
This is an attempt to keep the code base in compilable state, but
beyond that it is rather superficially

Gbp-Pq: Name 0012-Upstream-some-selected-bugfixes-from-cinelera-cv-unt.patch

+++ cinelerra/setformat.C
@@ -891,7 +891,7 @@ int ScaleAspectAuto::handle_event()
 	thread->update_aspect();
 }
 
-ScaleAspectText::ScaleAspectText(int x, int y, SetFormatThread *thread, float *output)
+ScaleAspectText::ScaleAspectText(int x, int y, SetFormatThread *thread, double *output)
  : BC_TextBox(x, y, 70, 1, *output)
 {
 	this->output = output;
